Inviter class moved to Inviter.cs
